ref: extract parse mod & add check mod
