+++ Cargo.toml
@@ -24,6 +24,7 @@ name = "bulloak"
 clap = { version = "4.3.19", features = ["derive"] }
 indexmap = "2.0.0"
 owo-colors = "3.5.0"
+solang-parser = "0.3.2"
 
 [dev-dependencies]
 pretty_assertions = { version= "1.4.0" }
