+++ src/lib.rs
@@ -1,0 +1,7 @@
+pub mod check;
+#[doc(hidden)]
+pub mod cli;
+pub mod parse;
+pub mod scaffold;
+#[doc(hidden)]
+mod utils;
